invitaPro/views/genericos.py
from django.shortcuts import render
from contextvars import Token

from myapps.authentication.manager import CustomUserManager
from myapps.authentication.models import UserCustomize as User
from django.http import JsonResponse
from django.utils.decorators import method_decorator  # importante
from django.shortcuts import render
from rest_framework.decorators import api_view, permission_classes
from rest_framework.permissions import AllowAny, IsAuthenticated
from rest_framework.response import Response
from rest_framework import generics, status
from django.conf import settings
from rest_framework.views import APIView
from myapps.authentication.permissions import HasRoleWithRoles
from myapps.authentication.authenticate import CustomJWTAuthentication
from django.utils import timezone
from django.db.models import Q
from django.db.models import Count
# Create your views here.


class RetriveTipoProducto(APIView):
    permission_classes = [AllowAny]
    
    def get(self, request):
        return Response("Ok", status=status.HTTP_200_OK)
